adding it all
