+++ package.json
@@ -78,7 +78,7 @@
     "prop-types": "^15.8.1",
     "react": "^18.1.0",
     "react-code-blocks": "^0.0.9-0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-leaflet": "^3.2.5",
     "react-leaflet-draw": "^0.19.8",
     "react-redux": "^7.2.6",
